JAVA/ChatGPT/Strategy/round-1/source/main/src/main/java/com/example/authentication/SessionAuthenticator.java
package com.example.authentication;


import javax.servlet.http.Cookie;
import javax.servlet.http.HttpServletRequest;
import java.util.HashMap;
import java.util.Map;
import java.util.UUID;

public class SessionAuthenticator implements Authenticator {
    private static final Map<String, String> sessions = new HashMap<>();

    public String createSession(String username) {
        String sessionId = UUID.randomUUID().toString();
        sessions.put(sessionId, username);
        return sessionId;
    }

    public boolean authenticate(HttpServletRequest request) {
        Cookie[] cookies = request.getCookies();
        if (cookies != null) {
            for (Cookie cookie : cookies) {
                if ("SESSIONID".equals(cookie.getName())) {
                    String sessionId = cookie.getValue();
                    return sessions.containsKey(sessionId);
                }
            }
        }
        return false;
    }
}